dials.array_family cleanup by namespace gatherer (#1162)
Move all code from flex.py to flex_ext.py
and turn flex.py into a pure importer.

This way there is no mangling of builtins in flex_ext.py,
and there is a full list of defined symbols in flex.py.

Even though LGTM complains about new alerts in the
pull request it will honour the suppression once merged.
